+++ darr/readcoderaggedarray.py
@@ -285,13 +285,15 @@ def readcode(dra, language, basepath='', abspath=False):
         vp = Path(dra._valuesdirname)
     return readcodefunc[language](dra=dra, indicespath=ip, valuespath=vp)
 
-shapeindexexplanationtextraggedarray = shapeexplanationtextarray + '\n' +  \
+
+shapeindexexplanationtextraggedarray = shapeexplanationtextarray + '\n\n' +  \
 wrap(f'Further, Python starts counting at 0. So the first subarray in a '
-     f'ragged array has index number 0. This is also true for IDL/GDL, but '
-     f'Julia, Mathematica, Matlab/Octave, R, and Maple start counting ' 
-     f'at 1, so the first subarray has index number 1 in these languages. '
-     f'Finally, in Python indexing the end index is non-inclusive. E.g., '
-     f'a[0:2] returns a[0] and a[1], but not a[2]. However, all other '
-     f'languages for which reading code is provided, Julia, Mathematica, '
-     f'Matlab/Octave, R, Maple, and IDL/GDL have an inclusive end index. The '
-     f'reading code provided takes these difference into account.\n')
+     f'ragged array has index number 0. This is also true for IDL/GDL. '
+     f'However, Julia, Mathematica, Matlab/Octave, R, and Maple start '
+     f'counting at 1, so the first subarray has index number 1 in these '
+     f'languages. Finally, in Python indexing the end index is '
+     f'non-inclusive. E.g., a[0:2] returns a[0] and a[1], but not a[2]. '
+     f'However, all other languages for which reading code is provided, Julia, '
+     f'Mathematica, Matlab/Octave, R, Maple, and IDL/GDL have an inclusive '
+     f'end index. The reading code provided takes these differences into '
+     f'account.\n')